+++ CHANGELOG.md
@@ -29,6 +29,12 @@ Test/interim releases are not documented here.
 
 ---
 
+## WIP
+### Changed (minor)
+- deprecate `RunnerSettings.diagramFile` and add `RunnerSettings.DiagramPath` as alternative.
+
+---
+
 ## [0.9.14-alpha]
 ### Fixed
 - Fix grammar - allow `via`, `exit`, `entry` to be used in user guard and action code